Finish the Rest APIs, Prepare to Learn WebSockets

+++ controllers/feed.js
@@ -2,34 +2,30 @@ const { validationResult } = require('express-validator/check');
 const fs = require('fs');
 const path = require('path');
 
-const Post  = require('../models/post')
+const Post  = require('../models/post');
+const User = require('../models/user');
 
 
-exports.getPosts = (req, res, next) => {
+exports.getPosts = async (req, res, next) => {
     const currentPage = req.query.page || 1;
     const perPage = 2;
     let totalItems;
-    Post.find().countDocuments()
-        .then(count => {
-            totalItems = count;
-            return Post.find()
-                .skip((currentPage - 1) * perPage)
-                .limit(perPage);
+    try {
+        const totalItems = await Post.find().countDocuments()
+        const posts = await Post.find()
+            .skip((currentPage - 1) * perPage)
+            .limit(perPage);
+        res.status(200).json({
+            message: 'Fetched posts successfully',
+            posts: posts,
+            totalItems: totalItems
         })
-        .then(posts => {
-            res.status(200).json({
-                message: 'Fetched posts successfully',
-                posts: posts,
-                totalItems: totalItems
-            })
-        })
-        .catch(err => {
-            if ( !err.statusCode ) {
-                err.statusCode = 500;
-            }
-            next(err)
-        })
-    
+    } catch(err) {
+        if ( !err.statusCode ) {
+            err.statusCode = 500;
+        }
+        next(err)
+    }
 };
 
 exports.createPost = (req, res, next) => {
@@ -44,23 +40,37 @@ exports.createPost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
     }
-    
+    console.log(req.file.path)
     const imageUrl = (req.file.path).replace('\\', '/');
     const title = req.body.title;
     const content = req.body.content
-    
+    let creator;
+    console.log(req.userId)
     const post = new Post({
         title: title, 
-        content: content,
         imageUrl: imageUrl,
-        creator: { name: 'drakeln' },
-    })
+        content: content,
+        creator: req.userId
+    });
     post.save()
     .then(result => {
-        console.log(result.imageUrl)
+        return User.findById(req.userId)
+    })
+    .then(user => {
+        console.log(user)
+        creator = user
+        console.log(post)
+        user.posts.push(post)
+        return user.save()
+    })
+    .then(result => {
         res.status(201).json({
             message: 'Post created successfully',
-            post: result
+            post: post,
+            creator: {
+                _id: creator._id,
+                name: creator.name
+            }
         })
     })
     .catch(err => {
@@ -120,6 +130,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error; // Catch will receive the errors
             }
+            if (post.creator.toString() !== req.userId ) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl ) {
                 clearImage(post.imageUrl)
             }
@@ -151,15 +166,29 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error; // Catch will receive the errors
             }
+            if (post.creator.toString() !== req.userId ) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl)
             return Post.findByIdAndRemove(postId)
         })
         .then(result => {
-            console.log(result);
-            res.status(200).json({message: ' Deleted'})
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId); // Removing postId exits into the user
+            return user.save()
+        })
+        .then(result => {
+            res.status(200).json({ message: 'Deleted post.' });
         })
         .catch(err => {
-
+            if (!err.statusCode) {
+                err.statusCode = 500;
+            }
+            next(err);
         })
 }
